requirements.txt
Django>=4.2,<5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2023.1
django-admin-tailwind==0.0.31
django-colorfield>=0.11.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
numpy>=1.21.0
numpy-stl>=2.17.0 